% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bmi.R
\name{calculate_bmi}
\alias{calculate_bmi}
\title{Calculates BMI}
\usage{
calculate_bmi(
  weight = NULL,
  height = NULL,
  weight_unit = "kg",
  height_unit = "cm"
)
}
\arguments{
\item{weight}{Weight of person}

\item{height}{Height of person}

\item{weight_unit}{The unit of weight parameter. Must be either 'kg' or 'lbs'.}

\item{height_unit}{The unit of the height parameter. Must be any of: 'cm', 'm', 'in' or 'ft'.}
}
\value{
A numerical value
}
\description{
Calculates BMI
}
\examples{
calculate_bmi(weight=80, height=180)
calculate_bmi(weight=160, height=6.4, weight_unit="lbs", height_unit="ft")
}
