% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bmi_prime.R
\name{calculate_bmi_prime}
\alias{calculate_bmi_prime}
\title{Calculates BMI Prime}
\usage{
calculate_bmi_prime(
  weight = NULL,
  height = NULL,
  weight_unit = "kg",
  height_unit = "cm",
  bmi = NULL,
  continent = NULL,
  upper_limit = NULL
)
}
\arguments{
\item{weight}{Weight of person}

\item{height}{Height of person}

\item{weight_unit}{The unit of weight parameter. Must be either 'kg' or 'lbs'.}

\item{height_unit}{The unit of the height parameter. Must be any of: 'cm', 'm', 'in' or 'ft'.}

\item{bmi}{BMI value. If provided, height and weight are not required.}

\item{continent}{Continent of person. Must be one of: 'Asia', 'Europe', 'North_America', 'South_America', 'Africa', 'Oceania'.}

\item{upper_limit}{Upper limit of BMI for the person's region. If provided, continent is not required.}
}
\value{
A numerical value
}
\description{
Calculates BMI Prime
}
\examples{
calculate_bmi_prime(bmi = 25, continent = "Europe")
calculate_bmi_prime(weight = 80, height = 180, continent = "Asia")
calculate_bmi_prime(bmi = 25, upper_limit = 25)
}
