% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bmi_prime.R
\name{calculate_bmi_prime}
\alias{calculate_bmi_prime}
\title{Calculates BMI prime}
\usage{
calculate_bmi_prime(
  height = NULL,
  weight = NULL,
  bmi = NULL,
  continent = NULL,
  upper_limit = NULL
)
}
\arguments{
\item{height}{Height in meters.}

\item{weight}{Weight in kilograms.}

\item{bmi}{Measured bmi as an alternative to supplying weight and height.}

\item{continent}{Continent that is to be used for normalisation.}

\item{upper_limit}{Alternative upper limit for specific country or state.}
}
\value{
A numerical value
}
\description{
Calculates BMI prime
}
\examples{
calculate_bmi_prime(height=1.80, weight=80, continent="Africa")
calculate_bmi_prime(bmi=22.9, continent="North_America")
calculate_bmi_prime(bmi=22.9, upper_limit=24.9)
}
